pyproject.toml
[tool.poetry]
name = "python_django_orm_blog"
version = "0.1.0"
description = "A simple Django project that implements a simple blog."
authors = ["Hexlet Team <info@hexlet.io>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.2.3"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.15.2"
pytest-django = "^4.3.0"
django-extensions = "^3.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
